Added comments and asserts

+++ week2/d02/Assignment_2_2_2/includes/utils.h
@@ -4,10 +4,10 @@
 #include <iostream>
 
 void print_help();
-void print_sum(int a, int b);
-void print_substraction(int a, int b);
-void print_multiplication(int a, int b);
-void print_division(int a, int b);
+void print_sum(double a, double b);
+void print_substraction(double a, double b);
+void print_multiplication(double a, double b);
+void print_division(double a, double b);
 void encrypt(std::string str, int key);
 void decrypt(std::string str, int key);
